+++ core/src/main/java/com/backblaze/b2/client/structures/B2Bucket.java
@@ -47,9 +47,6 @@ public class B2Bucket {
     @B2Json.optional
     private final B2AuthorizationFilteredResponseField<B2BucketReplicationConfiguration> replicationConfiguration;
 
-    @B2Json.optional
-    private final B2AuthorizationFilteredResponseField<List<B2EventNotificationRuleForResponse>> eventNotificationRules;
-
     @B2Json.required
     private final int revision;
 
@@ -65,7 +62,6 @@ public class B2Bucket {
                     B2AuthorizationFilteredResponseField<B2BucketFileLockConfiguration> fileLockConfiguration,
                     B2AuthorizationFilteredResponseField<B2BucketServerSideEncryption> defaultServerSideEncryption,
                     B2AuthorizationFilteredResponseField<B2BucketReplicationConfiguration> replicationConfiguration,
-                    B2AuthorizationFilteredResponseField<List<B2EventNotificationRuleForResponse>> eventNotificationRules,
                     int revision) {
         this.accountId = accountId;
         this.bucketId = bucketId;
@@ -78,7 +74,6 @@ public class B2Bucket {
         this.fileLockConfiguration = fileLockConfiguration;
         this.defaultServerSideEncryption = defaultServerSideEncryption;
         this.replicationConfiguration = replicationConfiguration;
-        this.eventNotificationRules = eventNotificationRules;
         this.revision = revision;
     }
 
@@ -187,38 +182,8 @@ public class B2Bucket {
         return replicationConfiguration.getValue();
     }
 
-    /**
-     * Indicates whether client is authorized to read event notification rule settings for bucket
-     * @return true iff client is authorized to read value of eventNotificationRules field in B2Bucket
-     */
-    public boolean isClientAuthorizedToReadEventNotificationRules() {
-        B2Preconditions.checkState(eventNotificationRules != null);
-
-        return eventNotificationRules.isClientAuthorizedToRead();
-    }
-
-    /**
-     * Returns settings for bucket event notification rules or null if there are none.
-     * Throws B2ForbiddenException if client is not authorized to read bucket event notification rule settings.
-     * @return event notification rule settings
-     * @throws B2ForbiddenException if client is not authorized to read eventNotificationRules field
-     */
-    public List<B2EventNotificationRuleForResponse> getEventNotificationRules() throws B2ForbiddenException {
-        B2Preconditions.checkState(eventNotificationRules != null);
-
-        // will throw B2ForbiddenException if client is not authorized to read value
-        return eventNotificationRules.getValue();
-    }
-
     @Override
     public String toString() {
-        String eventNotificationRulesDesc = "unauthorized to view";
-        try {
-            eventNotificationRulesDesc = String.valueOf(
-                    (eventNotificationRules == null ? 0 : eventNotificationRules.getValue().size())
-            );
-        } catch (B2ForbiddenException ignored) {}
-
         return "B2Bucket(" +
                 bucketName + "," +
                 bucketType + "," +
@@ -230,8 +195,7 @@ public class B2Bucket {
                 fileLockConfiguration + "," +
                 defaultServerSideEncryption + "," +
                 replicationConfiguration + "," +
-                eventNotificationRulesDesc + " eventNotificationRules," +
-                "v" + revision +
+                   "v" + revision +
                 ')';
     }
 
@@ -252,8 +216,7 @@ public class B2Bucket {
                 // don't use getter for these fields because they can throw B2ForbiddenException
                 Objects.equals(fileLockConfiguration, b2Bucket.fileLockConfiguration) &&
                 Objects.equals(defaultServerSideEncryption, b2Bucket.defaultServerSideEncryption) &&
-                Objects.equals(replicationConfiguration, b2Bucket.replicationConfiguration) &&
-                Objects.equals(eventNotificationRules, b2Bucket.eventNotificationRules);
+                Objects.equals(replicationConfiguration, b2Bucket.replicationConfiguration);
     }
 
     @Override
@@ -271,8 +234,7 @@ public class B2Bucket {
                 // don't use getter for these fields because they can throw B2ForbiddenException
                 fileLockConfiguration,
                 defaultServerSideEncryption,
-                replicationConfiguration,
-                eventNotificationRules
+                replicationConfiguration
         );
     }
 }