+++ core/src/main/java/com/backblaze/b2/client/structures/B2CreateBucketRequest.java
@@ -17,8 +17,6 @@ public class B2CreateBucketRequest {
     private final B2BucketServerSideEncryption defaultServerSideEncryption;
     private final B2BucketReplicationConfiguration replicationConfiguration;
 
-    private final List<B2EventNotificationRuleForRequest> eventNotificationRules;
-
     public B2CreateBucketRequest(String bucketName,
                                  String bucketType,
                                  Map<String, String> bucketInfo,
@@ -26,8 +24,7 @@ public class B2CreateBucketRequest {
                                  List<B2LifecycleRule> lifecycleRules,
                                  boolean fileLockEnabled,
                                  B2BucketServerSideEncryption defaultServerSideEncryption,
-                                 B2BucketReplicationConfiguration replicationConfiguration,
-                                 List<B2EventNotificationRuleForRequest> eventNotificationRules) {
+                                 B2BucketReplicationConfiguration replicationConfiguration) {
         this.bucketName = bucketName;
         this.bucketType = bucketType;
         this.bucketInfo = bucketInfo;
@@ -36,7 +33,6 @@ public class B2CreateBucketRequest {
         this.fileLockEnabled = fileLockEnabled;
         this.defaultServerSideEncryption = defaultServerSideEncryption;
         this.replicationConfiguration = replicationConfiguration;
-        this.eventNotificationRules = eventNotificationRules;
     }
 
     public String getBucketName() {
@@ -71,10 +67,6 @@ public class B2CreateBucketRequest {
         return replicationConfiguration;
     }
 
-    public List<B2EventNotificationRuleForRequest> getEventNotificationRules() {
-        return eventNotificationRules;
-    }
-
     public static Builder builder(String bucketName, String bucketType) {
         return new Builder(bucketName, bucketType);
     }
@@ -89,7 +81,6 @@ public class B2CreateBucketRequest {
         private boolean fileLockEnabled;
         private B2BucketServerSideEncryption defaultServerSideEncryption;
         private B2BucketReplicationConfiguration replicationConfiguration;
-        private List<B2EventNotificationRuleForRequest> eventNotificationRules;
 
         Builder(String bucketName,
                 String bucketType) {
@@ -127,11 +118,6 @@ public class B2CreateBucketRequest {
             return this;
         }
 
-        public Builder setEventNotificationRules(List<B2EventNotificationRuleForRequest> eventNotificationRules) {
-            this.eventNotificationRules = eventNotificationRules;
-            return this;
-        }
-
         public B2CreateBucketRequest build() {
             return new B2CreateBucketRequest(
                     bucketName,
@@ -141,8 +127,7 @@ public class B2CreateBucketRequest {
                     lifecycleRules,
                     fileLockEnabled,
                     defaultServerSideEncryption,
-                    replicationConfiguration,
-                    eventNotificationRules
+                    replicationConfiguration
             );
         }
     }