+++ core/src/test/java/com/backblaze/b2/client/B2StorageClientImplTest.java
@@ -21,8 +21,6 @@ import com.backblaze.b2.client.structures.B2CancelLargeFileResponse;
 import com.backblaze.b2.client.structures.B2CopyFileRequest;
 import com.backblaze.b2.client.structures.B2CreateBucketRequest;
 import com.backblaze.b2.client.structures.B2CreateBucketRequestReal;
-import com.backblaze.b2.client.structures.B2CustomHeaderForRequest;
-import com.backblaze.b2.client.structures.B2CustomHeaderForResponse;
 import com.backblaze.b2.client.structures.B2DeleteBucketRequest;
 import com.backblaze.b2.client.structures.B2DeleteBucketRequestReal;
 import com.backblaze.b2.client.structures.B2DeleteFileVersionRequest;
@@ -209,7 +207,6 @@ public class B2StorageClientImplTest extends B2BaseTest {
                 null,
                 null,
                 null,
-                null,
                 1
         );
         when(webifier.createBucket(anyObject(), anyObject())).thenReturn(bucket);
@@ -226,7 +223,6 @@ public class B2StorageClientImplTest extends B2BaseTest {
                         null,
                         null,
                         false,
-                        null,
                         null,
                         null
                 )
@@ -277,86 +273,12 @@ public class B2StorageClientImplTest extends B2BaseTest {
                         replicationRules,
                         sourceToDestinationKeyMapping
                 );
-        final List<B2EventNotificationRuleForRequest> eventNotificationRuleForRequests = listOf(
-                new B2EventNotificationRuleForRequest(
-                        "myRule",
-                        new TreeSet<>(
-                                listOf(
-                                        "b2:ObjectCreated:Replica",
-                                        "b2:ObjectCreated:Upload"
-                                )
-                        ),
-                        "",
-                        new B2WebhookConfigurationForRequest("https://www.example.com", null),
-                        true
-                ),
-                new B2EventNotificationRuleForRequest(
-                        "myRule2",
-                        new TreeSet<>(
-                                listOf(
-                                        "b2:ObjectDeleted:LifecycleRule"
-                                )
-                        ),
-                        "",
-                        new B2WebhookConfigurationForRequest(
-                                "https://www.example2.com",
-                                new TreeSet<>(
-                                        listOf(
-                                                new B2CustomHeaderForRequest("name1", "val1"),
-                                                new B2CustomHeaderForRequest("name2", "val2")
-                                        )
-                                )
-                        ),
-                        true
-                )
-        );
         final B2CreateBucketRequest request = B2CreateBucketRequest
                 .builder(BUCKET_NAME, BUCKET_TYPE)
                 .setBucketInfo(bucketInfo)
                 .setLifecycleRules(lifecycleRules)
                 .setDefaultServerSideEncryption(defaultServerSideEncryption)
-                .setEventNotificationRules(eventNotificationRuleForRequests)
                 .build();
-
-        final List<B2EventNotificationRuleForResponse> eventNotificationRuleForResponses = listOf(
-                new B2EventNotificationRuleForResponse(
-                        "myRule",
-                        new TreeSet<>(
-                                listOf(
-                                        "b2:ObjectCreated:Replica",
-                                        "b2:ObjectCreated:Upload"
-                                )
-                        ),
-                        "",
-                        new B2WebhookConfigurationForResponse(
-                                "https://www.example.com",
-                                "hmacSha256SigningSecret"
-                        ),
-                        true,
-                        false,
-                        ""),
-                new B2EventNotificationRuleForResponse(
-                        "myRule2",
-                        new TreeSet<>(
-                                listOf(
-                                        "b2:ObjectDeleted:LifecycleRule"
-                                )
-                        ),
-                        "",
-                        new B2WebhookConfigurationForResponse(
-                                "https://www.example2.com",
-                                new TreeSet<>(
-                                        listOf(
-                                                new B2CustomHeaderForResponse("name1", "val1"),
-                                                new B2CustomHeaderForResponse("name2", "val2")
-                                        )
-                                ),
-                                "hmacSha256SigningSecret"
-                        ),
-                        true,
-                        false,
-                        "")
-        );
 
         final B2Bucket bucket = new B2Bucket(
                 ACCOUNT_ID,
@@ -370,7 +292,6 @@ public class B2StorageClientImplTest extends B2BaseTest {
                 new B2AuthorizationFilteredResponseField<>(true, new B2BucketFileLockConfiguration(true)),
                 new B2AuthorizationFilteredResponseField<>(true, defaultServerSideEncryption),
                 new B2AuthorizationFilteredResponseField<>(true, replicationConfiguration),
-                new B2AuthorizationFilteredResponseField<>(true, eventNotificationRuleForResponses),
                 1
         );
         final B2CreateBucketRequestReal expectedRequest = new B2CreateBucketRequestReal(ACCOUNT_ID, request);
@@ -407,7 +328,7 @@ public class B2StorageClientImplTest extends B2BaseTest {
                 "isEnabled=true, includeExistingFiles=false}]}, " +
                 "asReplicationDestination=B2DestinationConfig{" +
                 "sourceToDestinationKeyMapping={123a0a1a2a3a4a50000bc614e=555a0a1a2a3a4a70000bc929a, " +
-                "456a0b9a8a7a6a50000fc614e=555a0a1a2a3a4a70000bc929a}}}}),2 eventNotificationRules,v1)",
+                "456a0b9a8a7a6a50000fc614e=555a0a1a2a3a4a70000bc929a}}}}),v1)",
                 bucket.toString()
         );
 
@@ -423,7 +344,6 @@ public class B2StorageClientImplTest extends B2BaseTest {
                 new B2AuthorizationFilteredResponseField<>(true, new B2BucketFileLockConfiguration(true)),
                 new B2AuthorizationFilteredResponseField<>(true, defaultServerSideEncryption),
                 new B2AuthorizationFilteredResponseField<>(true, replicationConfiguration),
-                new B2AuthorizationFilteredResponseField<>(true, eventNotificationRuleForResponses),
                 1
         );
 
@@ -443,7 +363,7 @@ public class B2StorageClientImplTest extends B2BaseTest {
                         "isEnabled=true, includeExistingFiles=false}]}, " +
                         "asReplicationDestination=B2DestinationConfig{" +
                         "sourceToDestinationKeyMapping={123a0a1a2a3a4a50000bc614e=555a0a1a2a3a4a70000bc929a, " +
-                        "456a0b9a8a7a6a50000fc614e=555a0a1a2a3a4a70000bc929a}}}}),2 eventNotificationRules,v1)",
+                        "456a0b9a8a7a6a50000fc614e=555a0a1a2a3a4a70000bc929a}}}}),v1)",
                 bucketWithOptions.toString());
     }
 
@@ -550,7 +470,6 @@ public class B2StorageClientImplTest extends B2BaseTest {
                 bucketId(i),
                 BUCKET_NAME,
                 BUCKET_TYPE,
-                null,
                 null,
                 null,
                 null,
@@ -681,7 +600,6 @@ public class B2StorageClientImplTest extends B2BaseTest {
                 .setBucketInfo(B2Collections.mapOf())
                 .setLifecycleRules(listOf())
                 .setBucketType(B2BucketTypes.ALL_PUBLIC)
-                .setEventNotificationRules(listOf())
                 .build();
         when(webifier.updateBucket(anyObject(), eq(request))).thenReturn(bucket(1));
 
@@ -698,7 +616,6 @@ public class B2StorageClientImplTest extends B2BaseTest {
                         .setBucketInfo(B2Collections.mapOf())
                         .setLifecycleRules(listOf())
                         .setBucketType(B2BucketTypes.ALL_PUBLIC)
-                        .setEventNotificationRules(listOf())
                         .build());
     }