+++ core/src/test/java/com/backblaze/b2/client/B2TestHelpers.java
@@ -175,7 +175,6 @@ public class B2TestHelpers {
                 null,
                 null,
                 null,
-                null,
                 i
         );
     }